+++ src/main/java/org/folio/rest/impl/InstanceBatchSyncUnsafeApi.java
@@ -8,10 +8,11 @@ import io.vertx.core.Handler;
 import java.util.Map;
 import javax.ws.rs.core.Response;
 import org.folio.rest.annotations.Validate;
+import org.folio.rest.jaxrs.model.Instances;
 import org.folio.rest.jaxrs.model.InstancesPost;
 import org.folio.rest.jaxrs.resource.InstanceStorageBatchSynchronousUnsafe;
 import org.folio.services.instance.InstanceService;
-import org.folio.utils.InstanceUtils;
+import org.folio.utils.ObjectConverterUtils;
 
 public class InstanceBatchSyncUnsafeApi implements InstanceStorageBatchSynchronousUnsafe {
   @Validate
@@ -20,7 +21,7 @@ public class InstanceBatchSyncUnsafeApi implements InstanceStorageBatchSynchrono
                                                         Handler<AsyncResult<Response>> asyncResultHandler,
                                                         Context vertxContext) {
 
-    var instances = InstanceUtils.copyPropertiesToInstances(entity.getInstances());
+    var instances = ObjectConverterUtils.convertObject(entity, Instances.class);
 
     new InstanceService(vertxContext, okapiHeaders)
       .createInstances(instances.getInstances(), true, false)