+++ src/main/java/org/folio/rest/impl/InventoryViewApi.java
@@ -1,7 +1,5 @@
 package org.folio.rest.impl;
 
-import static org.folio.rest.persist.PgUtil.streamGet;
-
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Context;
 import io.vertx.core.Handler;
@@ -9,8 +7,8 @@ import io.vertx.ext.web.RoutingContext;
 import java.util.Map;
 import javax.ws.rs.core.Response;
 import org.folio.rest.annotations.Validate;
-import org.folio.rest.jaxrs.model.InventoryViewInstance;
 import org.folio.rest.jaxrs.resource.InventoryViewInstances;
+import org.folio.services.instance.InstanceService;
 
 public class InventoryViewApi implements InventoryViewInstances {
   @Validate
@@ -19,7 +17,8 @@ public class InventoryViewApi implements InventoryViewInstances {
                                         RoutingContext routingContext, Map<String, String> okapiHeaders,
                                         Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
 
-    streamGet("instance_holdings_item_view", InventoryViewInstance.class, query,
-      offset, limit, null, "instances", routingContext, okapiHeaders, vertxContext);
+    new InstanceService(vertxContext, okapiHeaders)
+      .streamGetInventoryViewInstances("instance_holdings_item_view", query,
+        offset, limit, null, "instances", 0, routingContext);
   }
 }